requirements.txt
azure-ai-ml>=1.11.0
openai>=1.3.7
transformers>=4.38.0
torch>=2.0.0
datasets>=2.14.0
accelerate>=0.21.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
matplotlib>=3.7.0
azureml-core>=1.53.0
azureml-dataset-runtime>=1.53.0
python-dotenv>=1.0.0

# Additional dependencies for local A100 GPU training upcomment libraries
# nvidia-cuda-runtime-cu12
# nvidia-cudnn-cu12
# deepspeed>=0.10.0
# tensorboard>=2.12.0
# psutil>=5.9.0
# pynvml>=11.5.0